Hub: separate db connection from migrations
