+++ packages/hub/.gitignore
@@ -13,5 +13,3 @@
 build
 lib
 
-# This is checked in as a typescript file
-knexfile.js
