+++ packages/hub/.vscode/launch.json
@@ -12,7 +12,7 @@
         "node_modules/.bin/jest",
         "--runInBand",
         "--env=jsdom",
-        "--config=${workspaceRoot}/config/jest/jest.config.js",
+        "--config=${workspaceRoot}/jest/jest.config.js",
         "${relativeFile}"
       ],
       "cwd": "${workspaceRoot}",
@@ -31,7 +31,7 @@
         "node_modules/.bin/jest",
         "--runInBand",
         "--env=jsdom",
-        "--config=${workspaceRoot}/config/jest/jest.chain.config.js",
+        "--config=${workspaceRoot}/jest/jest.chain.config.js",
         "${relativeFile}"
       ],
       "cwd": "${workspaceRoot}",
