+++ packages/hub/jest/jest.config.js
@@ -1,11 +1,11 @@
 const {resolve} = require('path');
-const root = resolve(__dirname, '../../');
+const root = resolve(__dirname, '../');
 
 module.exports = {
   rootDir: root,
   collectCoverageFrom: ['src/**/*.{js,jsx,ts,tsx}'],
-  setupFiles: ['./config/env.js'],
-  setupFilesAfterEnv: ['./src/config/knexSetupTeardown.ts'],
+  setupFiles: ['./env.js'],
+  setupFilesAfterEnv: ['./jest/knex-setup-teardown.ts'],
   testMatch: ['<rootDir>/**/__test__/**/?(*.)(spec|test).ts'],
   testEnvironment: 'node',
   testURL: 'http://localhost',
