+++ packages/hub/package.json
@@ -9,14 +9,14 @@
   "scripts": {
     "db:drop": "dropdb hub_wallet_${NODE_ENV}",
     "db:create": "createdb hub_wallet_${NODE_ENV}",
-    "db:migrate": "npx knex migrate:latest --knexfile lib/config/knexfile",
-    "db:rollback": "npx knex migrate:rollback --knexfile lib/config/knexfile",
-    "db:seed": "npx knex seed:run --knexfile lib/config/knexfile",
-    "hub:start:dev": "node ./lib/start-dev-server.js",
+    "db:migrate": "npx knex migrate:latest --knexfile lib/db-admin/knexfile",
+    "db:rollback": "npx knex migrate:rollback --knexfile lib/db-admin/knexfile",
+    "db:seed": "npx knex seed:run --knexfile lib/db-admin/knexfile",
+    "hub:start:dev": "ts-node ./lib/start-dev-server.js",
     "hub:watch": "nodemon --watch 'src/**/*' -e ts,tsx --exec ts-node ./src/start-dev-server.ts",
     "firebase-relay:start:sdev": "nodemon --watch 'src/**/*' -e ts,tsx --exec ts-node src/message/firebase-relay",
-    "chain:test": "npx jest --runInBand -c ./config/jest/jest.chain.config.js",
-    "hub:test": "npx jest --runInBand -c ./config/jest/jest.config.js",
+    "chain:test": "npx jest --runInBand -c ./jest/jest.chain.config.js",
+    "hub:test": "npx jest --runInBand -c ./jest/jest.config.js",
     "test:ci": "run-s chain:test hub:test",
     "start:shared-ganache": "NODE_ENV=development npx start-shared-ganache",
     "build": "npx tsc",
@@ -38,7 +38,7 @@
     "dotenv-expand": "4.2.0",
     "firebase": "6.6.2",
     "fmg-core": "0.5.9",
-    "knex": "0.20.1",
+    "knex": "0.20.3",
     "koa": "2.11.0",
     "koa-body": "4.1.1",
     "koa-router": "7.4.0",
@@ -53,7 +53,6 @@
     "@types/eslint": "^6.1.3",
     "@types/eslint-plugin-prettier": "^2.2.0",
     "@types/jest": "24.0.21",
-    "@types/knex": "0.15.2",
     "@types/koa": "2.0.51",
     "@types/koa-bodyparser": "4.2.1",
     "@types/koa-router": "7.0.42",
