+++ packages/hub/tsconfig.json
@@ -10,6 +10,6 @@
     "esModuleInterop": true,
     "strict": false
   },
-  "include": ["src/**/*", "config/**/*"],
+  "include": ["src/**/*"],
   "exclude": ["**/*test_data.ts", "**/__mocks__/*"]
 }
